Moved api/node.js to index.js and adjusted associated file references
